+++ service/exceptions.py
@@ -211,3 +211,7 @@ class NonZeroDeploymentException(Exception):
 
 class TimeoutError(Exception):
     pass
+
+
+class BadInstanceCount(ServiceException):
+    pass